Only add beforeunload listener to stepper, dont use our own modal
